SW-4330-followup Add map style dropdown and fullscreen button to editable map view
- extracted out the map view style component for reuse
- added to the editable map view
- the color contrast can be off in outdoors view sometimes
